pyproject.toml
[tool.poetry]
name = "checkbox-sdk"
version = "1.0.0"
description = "Ukraine checkbox api python sdk"
authors = ["Roman Telezhynskyi <dismine@gmail.com>"]
maintainers = ["Roman Telezhynskyi <dismine@gmail.com>"]
repository = "https://github.com/dismine/checkbox-sdk"
readme = "README.md"
documentation = "https://checkbox-sdk.readthedocs.io"
license = "MIT"
packages = [{ include = "checkbox_sdk" }]
keywords = ["python", "sdk"]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.8"
PyJWT = "^2.4.0"
httpx = "^0.27.0"

# poetry add --group dev <package name>
[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
isort = "^5.10.1"
flake8 = "^5"
autoflake = "^1.4"
ipython = "^8.10.0"
pytz = "^2023.3"
sourcery = "^1.21.0"
pytest = "^8.3.1"
tox = "<4.12.0"
mypy = "^1.11.0"
coverage = "^7.6.0"
types-pyjwt = "^1.7.1"
pydantic = { extras = ["email"], version = "^2.8.2" }
pytest-asyncio = "^0.23.8"
python-magic = "^0.4.27"
ruff = "^0.6"
safety = "^3.2.6"

# poetry add --group docs <package name>
pylint = "^3.2.6"
[tool.poetry.group.docs.dependencies]
sphinx = "^7.1"
sphinx-prompt = "^1.5.0"

[tool.black]
line-length = 119
target-version = ['py310', 'py311']
exclude = '''
(
    \.eggs
  | \.git
  | build
  | dist
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119
known_third_party = ["httpx", "jwt", "httpcore"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
log_cli = true
log_cli_level = "CRITICAL"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
